src/requirements.txt
numpy==1.18.1
pandas==1.0.3
PyYAML==5.2
scipy==1.3.2
torch==1.4.0
tqdm==4.41.1